Changes following feedback
Prompt "Flight Center password" -> "Flight Password".

Fix segfault if pam conversation is not successful.  E.g., user ctrl-c
when entering password.

Set password so that it can be used by next authentication method.  This
allows the user to type their unix password at the "Flight Password"
prompt and be signed in without having to enter it a second time.  This
only works if `pam_flight` is before `pam_unix`, if they are the other
way around we ignore the previously entered password and prompt for a
new password.